+++ srs/src/compressedsrs.rs
@@ -1,0 +1,6 @@
+mod groupmap;
+
+impl CompressedSRS {
+    fn create(length: usize, num_y_coords: usize) -> Self {}
+    fn decompress(c: Self) -> Vec<(G::BaseField, G::BaseField)> {}
+}
